refactor(ui): decompose productivity screen God Class into view model and modular sub-components
